+++ packages/wrapper/CHANGELOG.md
@@ -14,7 +14,7 @@ All notable changes to npm module [monaco-editor-wrapper](https://www.npmjs.com/
 
 ## [6.2.0] - 2025-01-31
 
-- Update to monaco-vscode-api v13 [#836](https://github.com/TypeFox/monaco-languageclient/pull/829)
+- Update to monaco-vscode-api v13 [#836](https://github.com/TypeFox/monaco-languageclient/pull/836)
   - Updated all `@codingame/monaco-vscode` packages to `13.1.1`.
   - Breaking changes not in this API, but when using `@monaco-vscode-api`:
     - `@codingame/monaco-vscode-api` should not be installed as vscode anymore